+++ src/Pim/Bundle/DataGridBundle/Extension/Formatter/Property/ProductValue/BooleanProperty.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Pim\Bundle\DataGridBundle\Extension\Formatter\Property;
+namespace Pim\Bundle\DataGridBundle\Extension\Formatter\Property\ProductValue;
 
 /**
- * Flexible boolean field property, able to render boolean attribute type
+ * Boolean field property, able to render boolean attribute type
  *
  * @author    Filips Alpe <filips@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class FlexibleBooleanProperty extends FlexibleTwigProperty
+class BooleanProperty extends TwigProperty
 {
     /**
      * {@inheritdoc}