+++ src/Pim/Bundle/DataGridBundle/Extension/Formatter/Property/ProductValue/PriceProperty.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Pim\Bundle\DataGridBundle\Extension\Formatter\Property;
+namespace Pim\Bundle\DataGridBundle\Extension\Formatter\Property\ProductValue;
 
 /**
- * Flexible price field property, able to render price attribute type
+ * Price field property, able to render price attribute type
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class FlexiblePriceProperty extends FlexibleFieldProperty
+class PriceProperty extends FieldProperty
 {
     /**
      * {@inheritdoc}